infrastructure/pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = ""
authors = ["muhammad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.87.0"
constructs = "^10.2.69"
aws-cdk-aws-appsync-alpha = "^2.59.0a0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
